requirements.txt
setuptools>=65.0.0
wheel
numpy>=1.24.0
opencv-python>=4.8.0
pytesseract>=0.3.10
face-recognition>=1.3.0
mysql-connector-python>=8.2.0
Pillow>=10.0.0

